+++ packages/video_player/video_player/test/video_player_test.dart
@@ -1499,6 +1499,11 @@ class FakeVideoPlayerPlatform extends VideoPlayerPlatform {
   }
 
   @override
+  Future<void> setAllowBackgroundPlayback(bool allowBackgroundPlayback) async {
+    calls.add('setAllowBackgroundPlayback');
+  }
+
+  @override
   Widget buildView(int textureId) {
     return Texture(textureId: textureId);
   }