+++ packages/video_player/video_player_android/android/src/main/java/io/flutter/plugins/videoplayer/VideoPlayer.java
@@ -58,6 +58,15 @@ public abstract class VideoPlayer {
     exoPlayer.addListener(createExoPlayerEventListener(exoPlayer));
     setAudioAttributes(exoPlayer, options.mixWithOthers);
 
+    if (options.allowBackgroundPlayback) {
+        exoPlayer.setWakeMode(C.WAKE_MODE_NETWORK);
+        exoPlayer.setAudioAttributes(
+            new AudioAttributes.Builder()
+                .setUsage(C.USAGE_MEDIA)
+                .build(),
+            /* handleAudioFocus= */ !options.mixWithOthers);
+    }
+
     return exoPlayer;
   }