+++ packages/video_player/video_player_avfoundation/darwin/video_player_avfoundation/Sources/video_player_avfoundation/FVPVideoPlayerPlugin.m
@@ -322,4 +322,11 @@ static void upgradeAudioSessionCategory(AVAudioSessionCategory requestedCategory
 #endif
 }
 
+- (void)setAllowBackgroundPlayback:(BOOL)allowBackgroundPlayback 
+                            error:(FlutterError **)error {
+#if TARGET_OS_IOS
+  // Do nothing here, the background playback still works
+#endif
+}
+
 @end