+++ packages/video_player/video_player_avfoundation/pigeons/messages.dart
@@ -73,4 +73,6 @@ abstract class AVFoundationVideoPlayerApi {
   void pause(int playerId);
   @ObjCSelector('setMixWithOthers:')
   void setMixWithOthers(bool mixWithOthers);
+  @ObjCSelector('setAllowBackgroundPlayback:')
+  void setAllowBackgroundPlayback(bool allowBackgroundPlayback);
 }